requirements.txt
Flask>=3.0.0,<4.0.0
firebase-admin>=6.5.0,<7.0.0
python-dotenv>=1.0.1,<2.0.0
gunicorn>=23.0.0,<24.0.0
